# Conflicts:
#	com.unity.shadergraph/Editor/ShaderFoundry/BlockMerging/BlockMerger.cs
#	com.unity.shadergraph/Editor/ShaderFoundry/Linker/LegacyLinkerUtilities.cs
#	com.unity.shadergraph/Editor/ShaderFoundry/Provider/LegacyTemplateProvider.cs
#	com.unity.shadergraph/Editor/ShaderFoundry/Utilities/AttributeUtilities.cs

+++ com.unity.render-pipelines.universal/Editor/Decal/DecalShaderGraphGUI.cs
@@ -64,8 +64,8 @@ namespace UnityEditor.Rendering.Universal
 
         public DecalShaderGraphGUI()
         {
-            m_MaterialScopeList.RegisterHeaderScope(Styles.inputs, (uint)Expandable.Inputs, DrawExposedProperties);
-            m_MaterialScopeList.RegisterHeaderScope(Styles.advancedOptions, (uint)Expandable.Advanced, DrawSortingProperties);
+            m_MaterialScopeList.RegisterHeaderScope(Styles.inputs, Expandable.Inputs, DrawExposedProperties);
+            m_MaterialScopeList.RegisterHeaderScope(Styles.advancedOptions, Expandable.Advanced, DrawSortingProperties);
         }
 
         /// <summary>
